Rename aux.py to auxiliary.py
