+++ symbtools/auxiliary.py
